handle background permission denied error (#32)
Location permissions need to be set to always on for Fitdle app in order
to use the map in the background. This PR just makes sure that it is
properly set so google maps doesn't crash.
- If not set, prompt user to set it in settings
- Check again if set. If it still isn't show a dialogbox error telling
to them to set it and come back. Also pop back to dashboard.
- Change `avgPace` variable from double to num so the dashboard doesn't
crash when loading.
- don't pop and push when opening run screen from dashboard because the
dashboard screen is no longer named under routes inside main.dart